+++ src/contexts/useRsvpContext.ts
@@ -21,6 +21,7 @@ export type RsvpValue = Pick<
   | "handleDeleteRsvp"
   | "handleDeleteRsvpEntries"
   | "handleClearRsvpEntries"
+  | "liveAttendingCount"
 >;
 
 export const RsvpContext = createContext<RsvpValue | null>(null);